More "Wiki->Feature description->Solar" GH example files
With these added .gh scripts, the first draft of the "Solar" components is done, except for the "Solar Performance Ratio". which does not have a description on the Wiki yet. Maybe I (Amr) can try inserting this information? To learn how it's done?

Amr and Christoph should discuss the questions and new GitHub issues next week (Jan 25 - 29). After that, we can finalize this part :)